proper stack first line for FetchError instances
The `.stack` property gets cached in the `captureStackTrace()`
call, so whatever is set as the `name` and `message` at that time
will be used for the first line of the stack trace.

Before this patch, FetchError's stack would just say "Error"
as the first line. Now they correctly display the
"${name}: ${message}" of the error instances.

Test case included.