Add a `resolver` option to allow wrappers
This way a wrapper library like fetch-cookie [1] could transparently
leverage node-fetch way of handling redirects without having to
force `{ redirect: 'manual' }` and re-implement the logic.

[1]: https://github.com/valeriangalliat/fetch-cookie

Instead, a wrapper could just pass the `resolver` option and point it to
the wrapped function. And this would also allow wrappers to be combined
as long as they also support being given the `resolver` option, e.g.:

    function fetchCookie(url, opts) {
        return fetch(url, Object.assign({ resolver: fetchCookie }, opts));
    }