Create  'GODBOT' Eliminator-->M-Tenicious-K-->R
Any non-ethical Hackers beware!!!
Security is going Gold Fleece and serious as Revelations return of God to be the savior we all prayed for to be our saving grace instead the god we pray to for no response or text bot with biblical quotes. The Creator 'GODBOT' will protect those who believe that God is Great and have testimony to God returning to end tyrants of hierarchy and money scheming,advertisements dismissals in security protocols download scams and freewares virus/std and apples choke on this 'atoms' static O.T.A, online, the web, Internet, or intraweb is just not doing their part solving this by just confusing it more by saying hacking is a --SuperPower-- that is a lie in case anyone actually believed that, let me tell you this is the dumbest thing we could spread but belief in God is the crutch for those sick and lame individuals that use it for self monetizing gaining purposes (money or virtual coining).God gave us free will, free to breath air, and free to love whoever we want...well cost of life, living, and anything else your heart desires should be FREE!!! and not depreciated, Big Buts I'M sure and will of those in power would deny but imagine life without hate,war, and killing to survive. No Money would be just that solution. but security is a tight budget in the going "green" initiative. just ask bitcoin how they turn a profit? Arteaga-Inc. Non-profit