support `AbortSignal.reason`
In enviroments that implement it.

When an `AbortSignal#reason` has a value, i.e. in nodejs >= v17.2.0,
it is added on the returned `AbortError`. `AbortError#reason` would be
`undefined` otherwise.

Contary to the specification, when `AbortController#abort` is called,
with or without a reason, `fetch` _does not_ throw the
`AbortSignal#reason`.