fix: Redirect follows when using Request as input
Previously, there was a check in place that prevented a streaming
request body from being redirected. However, this check assumed that the
`fetch` function was invoked with the `fetch(url, initOptions)`
overload, as opposed to the `fetch(request)` signature.

This change makes the `instanceof` check against the (already cloned)
body in the follow request's init options.